gdt: Fix off by four bug in gdt_entry initialization
Somehow this managed to slip by until now (it finally blew in my face
when I started my attemps at implementing the IDT.

The flags_and_limit_hi_part field in struct gdt_entry is a 8-bit field
that holds a pair of 4-bit values: the highest four bits of the segment
limit and the segment flags.

Before, gdt_entry_create was trying to shove 16 bits of information in-
to this 8-bit field by storing the full byte of the limit_hi_part, plus
also shifting the flags out of rage (f << 8 instead of f << 4).

Signed-off-by: Vinícius Schütz Piva <vinicius.vsczpv@outlook.com>